pipeline/requirements.txt
# Core geospatial
rasterio==1.3.10
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
fiona==1.9.6

# STAC / remote data access
pystac-client==0.8.5
pystac==1.10.1
requests==2.32.3

# Raster analysis
numpy==1.26.4
rasterstats==0.19.0
scipy==1.13.0

# PostGIS / database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
GeoAlchemy2==0.15.1

# Pipeline utilities
python-dotenv==1.0.1
tqdm==4.66.4
structlog==24.2.0
click==8.1.7

# Testing
pytest==8.2.0
pytest-mock==3.14.0
